Add files via upload
This is the most stable and bug-free version of the script that I have developed. It still needs a few tweaks to clean up some of the logic that is creating a few errors. That said, the conversion tool will convert virtually any Gregorian calendar date between 1850 AD and 2150 AD into its equivalent Neoteran AR or ER date. As explained in the description, you need to have Python's Skyfield class libraries and JPL's de440s.bsp installed on your machine to run the script.